Pfiew this was long to fix

+++ 2D/Contact.cpp
@@ -12,14 +12,14 @@ int main(int argc, char **argv) {
 
     Image img(H.title, {n, n});
 
-    for (int x = 0; x < n; ++x)
-        for (int y = 0; y < n / 2; ++y) img.put({x, y}, WHITE);
+    for (size_t x = 0; x < n; ++x)
+        for (size_t y = 0; y < n / 2; ++y) img.put({int(x), int(y)}, WHITE);
     img.show();
 
-    for (int i = 0; i < 3 * n; i++) {
+    for (size_t i = 0; i < 3 * n; i++) {
         int total = 0;
-        for (int x = 1; x < n - 1; x++) {
-            for (int y = 1; y < n - 1; y++) {
+        for (size_t x = 1; x < n - 1; x++) {
+            for (size_t y = 1; y < n - 1; y++) {
                 if ((x + y + i) % 2 == 0) {
                     bool nb = false;
                     nb |= (img.at({x - 1, y}) == WHITE);
@@ -27,10 +27,10 @@ int main(int argc, char **argv) {
                     nb |= (img.at({x, y - 1}) == WHITE);
                     nb |= (img.at({x, y + 1}) == WHITE);
                     if (nb && prng.bernoulli(p)) {
-                        img.put({x, y}, WHITE);
+                        img.put({int(x), int(y)}, WHITE);
                         total++;
                     } else {
-                        img.put({x, y}, BLACK);
+                        img.put({int(x), int(y)}, BLACK);
                     }
                 }
             }
